Reorganize and clean up code. Get VUMPS to output real tensors for real inputs
