--- examples/contraction_sequence_optimization/results.txt
@@ -1,21 +1,0 @@
-
-# DimT = UInt64
-N  Fastest        Time ratio (ITensor/TensorOperations)
-4  Depth first    3.8570785167392287
-5  UInt16, UInt16 1.8563061055580312
-6  Depth first    0.5899351632071113
-7  UInt16, UInt16 0.24788804818923088
-8  UInt16, UInt16 0.2029114871802376
-9  UInt16, UInt16 0.13866186965147612
-10 UInt16, UInt16 0.09727571774556852
-
-# DimT = UInt128
-N  Fastest        Time ratio (ITensor/TensorOperations)
-4  Depth first    3.828718840927498
-5  Depth first    1.502178875285329
-6  Depth first    0.5906373865234624
-7  Depth first    0.20159976547940617
-8  UInt16, UInt16 0.12602799361705064
-9  UInt16, UInt16 0.08539257245368874
-10 UInt16, UInt16 0.058993701194375595
-
