--- examples/contraction_sequence_optimization/tmp.txt
@@ -1,20 +1,0 @@
-network =
-  [[11, 17], [2, 20], [15, 28, 38], [15, 21, 34],
-   [12, 21, 40], [10, 30], [16, 22, 32],
-   [13, 22, 38], [12, 20], [2, 27, 33],
-   [5, 16, 37], [3, 26, 40], [6, 24],
-   [14, 24, 36], [9, 18, 37], [1, 19, 31],
-   [8, 26], [11, 23], [6, 17, 39], [14, 19, 36],
-   [13, 30], [4, 25, 33], [7, 23, 39], [4, 28],
-   [5, 27, 32], [9, 25, 35], [8, 29], [10, 18, 31],
-   [7, 29, 35], [1, 3, 34]]
-
-stats_itensor.time = 23.597973444
-stats_itensor.value = (Any[Any[2, 9], Any[5, Any[4, Any[30, Any[Any[Any[3, 24], Any[8, Any[6, 21]]], Any[Any[Any[10, 22], Any[11, Any[7, 25]]], Any[Any[15, Any[28, Any[16, Any[20, Any[13, 14]]]]], Any[26, Any[Any[Any[1, 19], Any[18, 23]], Any[29, Any[27, Any[12, 17]]]]]]]]]]]], 91022703)
-
-stats_tensoroperations.time = 7.923444351
-stats_tensoroperations.value = (Any[Any[Any[13, Any[20, 14]], Any[19, Any[23, Any[1, 18]]]], Any[Any[29, Any[27, 17]], Any[16, Any[Any[Any[4, 30], Any[12, Any[5, Any[2, 9]]]], Any[Any[Any[22, 26], Any[24, 3]], Any[10, Any[Any[8, Any[11, Any[7, 25]]], Any[15, Any[28, Any[6, 21]]]]]]]]]], 1*χ^7 + 5*χ^6 + 8*χ^5 + 8*χ^4 + 6*χ^3 + 1*χ^2 + 0*χ + 0)
-
-stats_tensoroperations.time = 91.736996876
-stats_tensoroperations.value = (Any[Any[Any[13, Any[20, 14]], Any[19, Any[23, Any[1, 18]]]], Any[Any[29, Any[27, 17]], Any[16, Any[Any[Any[4, 30], Any[12, Any[5, Any[2, 9]]]], Any[Any[Any[22, 26], Any[24, 3]], Any[10, Any[Any[8, Any[11, Any[7, 25]]], Any[15, Any[28, Any[6, 21]]]]]]]]]], 884)
-
